Add files via upload
This paper introduces and outlines the foundations of the Generalized and Scalable Optimal Sparse Decision Trees (GOSDT). The paper offers insights into the optimization process for creating interpretable decision trees. 

In the file named OptimalTreesTf_IDF I implemented the GOSDT model with TF-IDF feature representation, achieving 100% training accuracy on my dataset. The dataset is named Womens_Clothing_E-Commerce_Reviews, and a more detailed description of this dataset can be also found in this repository.